feat: 관리자만 강사 crud 가 가능하도록 수정
- 강사 등록, 조회는 관리자만 가능
- 강사 수정, 삭제는 MANAGER 만 가능

+++ src/main/java/com/example/lecturerestapi/service/TeacherService.java
@@ -2,6 +2,8 @@ package com.example.lecturerestapi.service;
 
 import com.example.lecturerestapi.dto.TeacherRequest;
 import com.example.lecturerestapi.dto.TeacherResponse;
+import com.example.lecturerestapi.entity.Admin;
+import com.example.lecturerestapi.entity.AdminRole;
 import com.example.lecturerestapi.entity.Teacher;
 import com.example.lecturerestapi.repository.TeacherRepository;
 import lombok.RequiredArgsConstructor;
@@ -13,26 +15,37 @@ import org.springframework.transaction.annotation.Transactional;
 public class TeacherService {
     private final TeacherRepository teacherRepository;
 
-    public TeacherResponse save(TeacherRequest request) {
-        return new TeacherResponse(teacherRepository.save(request.toEntity()));
+    public TeacherResponse save(TeacherRequest request, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER || admin.getRole() == AdminRole.STAFF)
+            return new TeacherResponse(teacherRepository.save(request.toEntity()));
+        throw new IllegalArgumentException("관리자만 강사 등록이 가능합니다.");
     }
 
     @Transactional
-    public TeacherResponse update(Long teacherId, TeacherRequest request) {
-        Teacher teacher = findTeacherById(teacherId);
-        teacher.update(request);
-        return new TeacherResponse(teacher);
+    public TeacherResponse update(Long teacherId, TeacherRequest request, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER) {
+            Teacher teacher = findTeacherById(teacherId);
+            teacher.update(request);
+            return new TeacherResponse(teacher);
+        }
+        throw new IllegalArgumentException("MANAGER 만 강사 수정이 가능합니다.");
     }
 
-    public TeacherResponse findById(Long teacherId) {
-        Teacher teacher = findTeacherById(teacherId);
-        return new TeacherResponse(teacher);
+    public TeacherResponse findById(Long teacherId, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER || admin.getRole() == AdminRole.STAFF) {
+            Teacher teacher = findTeacherById(teacherId);
+            return new TeacherResponse(teacher);
+        }
+        throw new IllegalArgumentException("관리자만 강사 조회가 가능합니다.");
     }
 
-    public Long delete(Long teacherId) {
-        Teacher teacher = findTeacherById(teacherId);
-        teacherRepository.deleteById(teacher.getId());
-        return teacher.getId();
+    public Long delete(Long teacherId, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER) {
+            Teacher teacher = findTeacherById(teacherId);
+            teacherRepository.deleteById(teacher.getId());
+            return teacher.getId();
+        }
+        throw new IllegalArgumentException("MANAGER 만 강사 삭제가 가능합니다.");
     }
 
     public Teacher findTeacherById(Long teacherId) {
